New improved defrag algorithm. Now the fragmentation is optimized since the record is oversized. Oversize is configurable per class. Created new DEFRAG strategy global configuration property.

+++ core/src/main/java/com/orientechnologies/orient/core/record/ORecord.java
@@ -206,5 +206,5 @@ public interface ORecord<T> extends ORecordElement, OIdentifiable {
 	 * 
 	 * @return
 	 */
-	public long getSize();
+	public int getSize();
 }
